Externalize syncable.connect
